tools/lightroom-coord-picker/requirements.txt
pyautogui>=0.9.54
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
